Add dedicated shader for colored circle overlay
This lets us put the circle overlay on top of the blob shader, making
the blob itself not see and reflect the overlay.
Because of the change there is currently no vignette on top of the
overlay, but it should be possible to move some vignettes around to get
this back if we want to.